requirements.txt
requests
streamlit
pycountry
python-dotenv